Move RevHub to internal class
