+++ .gitignore
@@ -38,7 +38,7 @@ bower_components
 # node-waf configuration
 .lock-wscript
 
-# Compiled binary addons (https://nodejs.org/api/addons.html)
+# Compiled binary internal (https://nodejs.org/api/addons.html)
 build/Release
 
 # Dependency directories
