+++ RHSPlib/arch/linux/RHSPlib_serial.c
