+++ RHSPlib/rev/RHSPlib.c
