+++ RHSPlib/rev/RHSPlib_device_control.c
