+++ RHSPlib/rev/RHSPlib_dio.c
