+++ RHSPlib/rev/RHSPlib_errors.h
