+++ RHSPlib/rev/RHSPlib_motor.c
