+++ src/RHSPlibWorker.cc
