+++ src/RHSPlibWorker.h
