+++ src/addon.cc
